Mainly, update romonitor to use the idle subsystem, which will prevent the linux kernel from thinking it's performing some kind of attack

+++ update_albumart.php
@@ -9,23 +9,27 @@ switch (ROMPR_IMAGE_VERSION) {
     case 4:
         $na =  generic_sql_query("SELECT COUNT(Albumindex) AS NumAlbums FROM Albumtable WHERE Image LIKE 'albumart/small/%'", false, null, 'NumAlbums', 0);
         debuglog("There are ".$na." albums","AA_UPGRADE");
-        
+
         $k = generic_sql_query("SELECT ImgKey FROM Albumtable WHERE Image LIKE 'albumart/small/%' AND ImgVersion < ".ROMPR_IMAGE_VERSION." LIMIT 1", false, null, 'ImgKey', null);
         if ($k) {
             // We're dealing with a specific version here, where all images are jpg.
             // We'll mimic the behaviour of baseAlbumImage at the point in time, so if it changes
             // in the future it doesn't mess this up.
             $source = "albumart/asdownloaded/".$k.".jpg";
-            debuglog("Converting image ".$k,"AA_UPGRADE");
-            $ih = new imageHandler($source);
-            $ih->resizeToWidth(400);
-            $ih->save("albumart/medium/".$k.".jpg", 70);
-            $ih->resizeToWidth(100);
-            $ih->save("albumart/small/".$k.".jpg", 75);
-            $ih->destroy();
-            generic_sql_query("UPDATE Albumtable SET ImgVersion = ".ROMPR_IMAGE_VERSION." WHERE ImgKey = '".$k."'");
+            if (file_exists($source)) {
+                debuglog("Converting image ".$k,"AA_UPGRADE");
+                $ih = new imageHandler($source);
+                $ih->resizeToWidth(400);
+                $ih->save("albumart/medium/".$k.".jpg", 70);
+                $ih->resizeToWidth(100);
+                $ih->save("albumart/small/".$k.".jpg", 75);
+                $ih->destroy();
+                generic_sql_query("UPDATE Albumtable SET ImgVersion = ".ROMPR_IMAGE_VERSION." WHERE ImgKey = '".$k."'");
+            } else {
+                generic_sql_query("UPDATE Albumtable SET Image = '' WHERE ImgKey = '".$k."'");
+            }
         }
-        
+
         $oa =  generic_sql_query("SELECT COUNT(ImgVersion) AS NumOldAlbums FROM Albumtable WHERE Image LIKE 'albumart/small/%' AND ImgVersion < ".ROMPR_IMAGE_VERSION, false, null, 'NumOldAlbums', 0);
         debuglog("There are ".$oa." albums with old-style album art","AA_UPGRADE");
         break;
